Refactor: Update GameTimer models to use Int for timer values

+++ GameTime/Features/GameTimer/Models/GameTimer.Draft.swift
@@ -4,7 +4,7 @@ extension GameTimer {
     struct Draft: Equatable {
         let id: UUID
         var name: String?
-        var timer: Double?
+        var timer: Int?
     }
 }
